Make sure release is just one file
