+++ src/util/fs.js
@@ -513,9 +513,36 @@ export function copy(src: string, dest: string, reporter: Reporter): Promise<voi
   return copyBulk([{src, dest}], reporter);
 }
 
+function WorkerScript() {
+  // @flow
+  const {parentPort} = require('worker_threads');
+  const fs = require('fs');
+
+  parentPort.on('message', o => {
+    try {
+      let running = o.actions.length;
+      // Safety short circuit in case we somehow start a worker with nothing.
+      running === 0 && parentPort.postMessage('');
+
+      o.actions.forEach(a => {
+        fs.copyFile(a.src, a.dest, 0, err => {
+          if (err) {
+            parentPort.emit('error', err);
+          } else {
+            running -= 1;
+            running === 0 && parentPort.postMessage('');
+          }
+        });
+      });
+    } catch (e) {
+      parentPort.emit('error', e);
+    }
+  });
+}
+
 const {Worker} = require('worker_threads');
 function spawnWorker(): Worker {
-  return new Worker(require('path').join(__dirname, '..', 'worker.js'));
+  return new Worker(`(${WorkerScript.toString()})()`, {eval: true});
 }
 
 const numberOfWorkers = process.env.WORKERS_LIMIT
